[WIP] Add fx.OnStart
fx.OnStart reduces boilerplate when adding a function to be run
when the application is started.

It reduces boilerplate like:
```go
fx.Invoke(
    func(lc fx.Lifecycle, a *Type1, b *Type2) {
        lc.Append(fx.Hook{
            OnStart: func(ctx context.Context) error {
                doSomething(a, b)
            },
        })
    },
)
```

into

```go
fx.OnStart(
    func(a *Type1, b *Type2) {
        doSomething(a, b)
    },
)
```

This is a WIP implementation which implements the logic "outside" of fx, to start
a discussion of whether this is what we want.

There are some details to consider:
 * What happens if the graph has a context and the passed in function has a context
   Currently: If the first argument is a context, it's always the Start context. Any
   other argument as context is provided from the graph.
 * What is the source of the OnStart for errors?
   Today it'll be a generated function, but fx should probably skip these.
   We might want a `t.Helper()` to provide fx helpers that don't pollute the
   stack.

This could also be implemented "internally" in fx, which would complicate the core
framework but possibly provide better stacks/errors.

See T2133659 for more details.